# doc/api.yaml
openapi: 3.0.0
paths:
...
  /track:
    get:
      description: Retrieve a list of all tracks
      operationId: TracksController_getAllTracks
      parameters: []
      responses:
        "200":
          description: List of tracks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
      security: &a5
        - bearer: []
      summary: Get all tracks
      tags: &a6
        - Tracks
    post:
      description: Create a new track with name, duration, and optional artist/album
        references
      operationId: TracksController_createTrack
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "201":
          description: Track created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request - invalid input data
      security: *a5
      summary: Create a new track
      tags: *a6
  /track/{id}:
    get:
      description: Retrieve a specific track by its ID
      operationId: TracksController_getTrackById
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "200":
          description: Track retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found
      security: *a5
      summary: Get track by ID
      tags: *a6
    put:
      description: Update track information
      operationId: TracksController_updateTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: Track updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request - invalid input data
        "404":
          description: Track not found
      security: *a5
      summary: Update track
      tags: *a6
    delete:
      description: Delete a track by its ID. This will also remove the track from
        favorites if present.
      operationId: TracksController_deleteTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Track deleted successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found
      security: *a5
      summary: Delete track
      tags: *a6
...
  /favs/track/{id}:
    post:
      description: Add a track to favorites by its ID
      operationId: FavoritesController_addTrackToFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "201":
          description: Track added to favorites successfully
        "400":
          description: Invalid UUID format
        "422":
          description: Track not found - cannot add to favorites
      security: *a8
      summary: Add track to favorites
      tags: *a9
    delete:
      description: Remove a track from favorites by its ID
      operationId: FavoritesController_removeTrackFromFavorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        "204":
          description: Track removed from favorites successfully
        "400":
          description: Invalid UUID format
        "404":
          description: Track not found in favorites
      security: *a8
      summary: Remove track from favorites
      tags: *a9
...
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: string
          description: Track UUID
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: object
          description: Artist ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        albumId:
          type: object
          description: Album ID
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 354
      required:
        - id
        - name
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: object
          description: Artist ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        albumId:
          type: object
          description: Album ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 354
          minimum: 1
      required:
        - name
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: object
          description: Artist ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        albumId:
          type: object
          description: Album ID (optional)
          example: 123e4567-e89b-12d3-a456-426614174000
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 354
          minimum: 1
...